Add Nif derivication encodings for types that can be derieved
A couple of strategies were chosen to encode terms.

Structs are often given for Taiga terms, these correspond to maps in
BEAM languages.

Further any struct that is just a struct over an Array got a tuple
specification as they are just wrapper types

This commit Tries to minimze the amount of manual derived terms,
however they do appear in trying to derive some other terms. For these
types, tuples ewre used to make easy implemenattions of serialize and
desearlize